// src/tmdb/schema.d.ts
export interface operations {
// ...
    "discover-movie": {
// ...
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /8YFL5QQVPy3AgrEQxNYVSgiPEbe.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 640146
                             */
                            id: number;
                            /** @example en */
                            original_language?: string;
                            /** @example Ant-Man and the Wasp: Quantumania */
                            original_title?: string;
                            /** @example Super-Hero partners Scott Lang and Hope van Dyne, along with with Hope's parents Janet van Dyne and Hank Pym, and Scott's daughter Cassie Lang, find themselves exploring the Quantum Realm, interacting with strange new creatures and embarking on an adventure that will push them beyond the limits of what they thought possible. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 9272.643
                             */
                            popularity: number;
                            /** @example /ngl2FKBlU4fhbdsrtdom9LVLBXw.jpg */
                            poster_path?: string;
                            /** @example 2023-02-15 */
                            release_date?: string;
                            /** @example Ant-Man and the Wasp: Quantumania */
                            title?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 6.5
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 1856
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 38020
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 760385
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
}